try to make this faster and use less queries: o add hints to a few of the more general functions to keep the number of queries down o extend queries so that we can pass along the hints o start specifying columns instead of using *: since Database make the associated array, this should help speed things up alot
A few notes:
o I need to read up on using * as opposed to column names to see what the actual speed increase is (but since the php has to do tons less, it still might be better even if the sql stuff isn't any better
o There is a possibility that doing queries on multiple tables at once with larger where blocks is actually much slower, in which case we may go back to the high amout of very small and fast queries. Again, I will read the MySQL docs on query optimization.

In any case, this is a step in the right direction.


git-svn-id: file:///home/mjibson/backup/arcrepo/trunk@362 157e5a38-50d5-0310-a571-dcc21a78b580