Disable cache control
I published a page, viewed it, made a change and republished. The second version is viewable through the cloud storage console, but the old version still returns via the published URL. I'm making a random guess that somewhere between here and there something is caching that page, as it defaults to one hour. This could potentially happen to a user, so try disabling caching until I know more.