Support line and text editing with hallo
Separate lines and text fields. This will provide more granular difference between things that can support newlines and those that can't. This is important because there's no good way to limit a contentEditable div to a single line.

Hallo works, but is not a very user friendly solution. Will replace with something else soon.