Use a (possibly) singleton for global access
Instead of hanging everything off the MainActivity which could die, try
this. I'm aware that subclassing application is supposed to be the best
way, but this works just fine and is easier. There's a chance that it has
some horrible bug I'm not aware of, but even with background activities
off it worked.